+++ src/main/java/dwight/Launcher.java
@@ -1,10 +1,10 @@
-package duke;
+package dwight;
 
 import javafx.application.Application;
 
 public class Launcher {
 
     public static void main(String[] args) {
-        Application.launch(Duke.class, args);
+        Application.launch(Dwight.class, args);
     }
 }
